Adding recursive deletion and prompting for directories

+++ src/src/main/java/org/nextbox/service/FilesystemAPI.java
@@ -1,6 +1,7 @@
 package org.nextbox.service;
 
 
+import org.apache.commons.io.FileUtils;
 import org.nextbox.model.Filepath;
 import org.nextbox.model.User;
 import org.nextbox.model.Directory;
@@ -9,6 +10,7 @@ import org.nextbox.model.File;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.PrintStream;
 import java.nio.file.*;
 import java.util.ArrayList;
 import javax.swing.JOptionPane;
@@ -84,19 +86,37 @@ public class FilesystemAPI {
         int choice = JOptionPane.showConfirmDialog(null, "Really delete this?", "Warning", JOptionPane.YES_NO_OPTION);
         if (choice == JOptionPane.YES_OPTION)
         {
-            // Try block taken from Oracle File class documentation: https://docs.oracle.com/javase/tutorial/essential/io/delete.html
-            try {
-                Files.delete(toDelete.getPath().toAbsolutePath());
-            } catch (NoSuchFileException x) {
-                System.err.format("%s: no such" + " file or directory%n", toDelete.getPath().toAbsolutePath());
-                return false;
-            } catch (DirectoryNotEmptyException x) {
-                System.err.format("%s not empty%n", toDelete.getPath().toAbsolutePath());
-                return false;
-            } catch (IOException x) {
-                // File permission problems are caught here.
-                System.err.println(x);
-                return false;
+            if (toDelete.pathIsDir()) {
+                // Try block taken from Oracle File class documentation: https://docs.oracle.com/javase/tutorial/essential/io/delete.html
+                try {
+                    Files.delete(toDelete.getPath().toAbsolutePath());
+                } catch (NoSuchFileException x) {
+                    System.err.format("%s: no such" + " file or directory%n", toDelete.getPath().toAbsolutePath());
+                    return false;
+                } catch (DirectoryNotEmptyException y) {
+                    System.err.format("%s not empty%n", toDelete.getPath().toAbsolutePath());
+                    int recurse = JOptionPane.showConfirmDialog(null, "Recursively delete directory?", "Warning", JOptionPane.YES_NO_OPTION);
+                    if (recurse == JOptionPane.YES_OPTION) {
+                        try {
+                            FileUtils.deleteDirectory(toDelete.fptoFile());
+                        } catch (NoSuchFileException z) {
+                            System.err.format("%s: no such" + " file or directory%n", toDelete.getPath().toAbsolutePath());
+                            return false;
+                        } catch (IOException w) {
+                            // File permission problems are caught here.
+                            System.err.println(w);
+                            return false;
+                        }
+                    }
+                    else {
+                        return true;
+                    }
+                } catch (IOException x) {
+                    // File permission problems are caught here.
+                    System.err.println(x);
+                    return false;
+                }
+                return true;
             }
             return true;
         }
